moved functionality and broke it into files
