+++ client/index.android.js
@@ -5,148 +5,19 @@
  */
 
 import React, {Component} from "react";
-import {AppRegistry, Button, ListView, ScrollView, StyleSheet, Text, TextInput, View} from "react-native";
-import {createStore} from "redux";
+import {AppRegistry} from "react-native";
 import Provider from "react-redux/es/components/Provider";
-import connect from "react-redux/es/connect/connect";
-import * as AsyncStorage from "react-native/Libraries/Storage/AsyncStorage";
-
-const actions = {
-	addMessage: function (state, action) {
-		AsyncStorage.setItem('@Storage:messages', JSON.stringify([...state.messages, action.message]));
-		return {
-			...state,
-			messages: [...state.messages, action.message]
-		}
-	},
-	setInputText: function (state, action) {
-		return {
-			...state,
-			inputText: action.text
-		}
-	},
-	setMessages: function (state, action) {
-		return {
-			...state,
-			messages: action.messages
-		}
-	}
-};
-
-
-AsyncStorage.getItem('@Storage:messages').then(function (messages) {
-	store.dispatch({
-		type: 'setMessages',
-		messages: JSON.parse(messages) || []
-	});
-});
-
-const initialState = {
-	inputText: '',
-	messages: [],
-};
-
-function messenger(state = initialState, action) {
-	const operation = actions[action.type];
-	return operation ? operation(state, action) : state;
-}
-
-function mapStateToProps(state) {
-	return state;
-}
-
-const store = createStore(messenger);
+import {Messenger} from "./components/messenger";
+import {MessengerStore} from "./utils/redux-stores";
 
 export default class E2essage extends Component {
 	render() {
 		return (
-			<Provider store={store}>
+			<Provider store={MessengerStore}>
 				<Messenger/>
 			</Provider>
 		);
 	}
 }
-
-class Mess extends Component {
-	constructor(props) {
-		super(props);
-		this.ds = new ListView.DataSource({rowHasChanged: (r1, r2) => r1.text !== r2.text});
-	};
-
-	addMessage(text) {
-		store.dispatch({
-			type: 'addMessage',
-			message: {
-				text,
-				style: styles.bigRed
-			}
-		});
-		this.refs.messageInput.setNativeProps({text: ''});
-		this.setInputText('');
-	}
-
-	setInputText(text) {
-		store.dispatch({
-			type: 'setInputText',
-			text
-		})
-	}
-
-	render() {
-		return (
-			<View style={{
-				flex: 1,
-				flexDirection: 'column',
-			}}>
-				<Text style={styles.bigRed}>Messages</Text>
-				<ScrollView
-					ref='messages'
-					keyboardShouldPersistTaps='always'
-					onContentSizeChange={(contentWidth, contentHeight) =>
-					{this.refs.messages.scrollTo({y: contentHeight});
-					}}
-				>
-					<ListView
-						enableEmptySections={true}
-						styles={{flex: 1}}
-						dataSource={this.ds.cloneWithRows(this.props.messages)}
-						renderRow={(row) => <Text style={row.style}>{row.text}</Text>}
-					/>
-				</ScrollView>
-				<View style={{
-					flexDirection: 'row',
-				}}>
-					<TextInput
-						ref="messageInput"
-						style={{height: 40, flex: 3}}
-						placeholder="What do you want to say?"
-						onChangeText={(text) => this.setInputText(text)}
-					/>
-					<Button
-						style={{height: 40, flex: 1}}
-						title="Send"
-						onPress={() => {
-							this.props.inputText && this.addMessage(this.props.inputText);
-						}}
-					/>
-				</View>
-			</View>
-		)
-	}
-}
-
-const Messenger = connect(mapStateToProps)(Mess);
-
-const styles = StyleSheet.create({
-	bigRed: {
-		color: 'red',
-		fontWeight: 'bold',
-		fontSize: 30,
-	},
-	blue: {
-		color: 'blue',
-	}
-});
-
 
 AppRegistry.registerComponent('client', () => E2essage);
