+++ client/utils/create-store.js
@@ -1,0 +1,9 @@
+import {createStore} from "redux";
+
+
+export function createStoreWithActions(initialState, actionsObject) {
+	return createStore(function (state = initialState, action) {
+		const operation = actionsObject[action.type];
+		return operation ? operation(state, action) : state;
+	});
+}
